Guidance on sequence number ranges to report upon
Last paragraph of Section 3 is added, giving guidance on what sequence
number range should be included in each report. State that they should
cover consecutive ranges, and not overlap, but state behaviour if they
do overlap. State that reports more than one quarter of the sequence
number space ahead or behind the previous report MUST be ignored.

This is a technical change to the specification.